+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Kube Utils Changelog
 
+## [21.1.1]
+
+### Added 
+
+- `findFirstBlockTagWithinRadius` to find a block within a radius that has a specific block tag.
+- `findAnyEntitiesWithinRadius` to find any entities within a radius based on an entity type.
+- `findLivingEntitiesWithinRadius` to find living entities within a radius based on an entity type.
+
+### Deprecated
+
+- `findEntitiesWithinRadius` in favour of `findLivingEntitiesWithinRadius`
+
 ## [21.1.0]
 
 ### Changed